document_processing_assistant/requirements.txt
# streamlit==1.26.0
streamlit
pandas==2.0.3
# pillow==10.0.0
pillow
pymupdf==1.22.5
markdown==3.4.4
pytest==7.4.0
numpy==1.25.2